// ai-demo/app/form/types.ts
export interface ShopInfoData {
  shopName: string
  industry: string
  description: string
  established: number | null
  prefecture: string
  city: string
  streetAddress: string
  phone: string | null
  email: string | null
  openingTime: string
  closingTime: string
  regularHoliday: string | null
  parking: string | null
  websiteUrl: string | null
  instagramUrl: string | null
  xUrl: string | null
  announcement: string | null
}